:bug: fix helper
